# docs/swagger.yaml
basePath: /api/
definitions:
...
  messages.IdResponse:
    properties:
      id:
        type: string
    type: object
  messages.MessageListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  messages.MessageResponse:
    properties:
      message:
        $ref: '#/definitions/models.Message'
    type: object
  messages.TextInput:
    properties:
      text:
        type: string
    type: object
...
  models.Message:
    properties:
      author:
        type: integer
      chat_id:
        type: integer
      id:
        type: integer
      sent_at:
        type: string
      text:
        type: string
    required:
    - text
    type: object
  models.User:
    properties:
      icon:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - icon
    - password
    - username
    type: object
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: Tairyoku
    url: https://github.com/Tairyoku
  description: Серверна частина онлайн чату.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Server API
  version: 1.0.0
paths:
...
  /chats/{chatId}/messages:
    post:
      consumes:
      - application/json
      description: |-
        Отримує текст повідомлення.
        Створює повідомлення.
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: Message text
        in: body
        name: message_text
        required: true
        schema:
          $ref: '#/definitions/messages.TextInput'
      produces:
      - application/json
      responses:
        "200":
          description: return message ID
          schema:
            $ref: '#/definitions/messages.IdResponse'
        "400":
          description: body is empty
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: create message error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create message
      tags:
      - message
  /chats/{chatId}/messages/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID повідомлення.
        Повертає повідомлення.
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: return message ID
          schema:
            $ref: '#/definitions/messages.MessageResponse'
        "500":
          description: get message error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get message by ID
      tags:
      - message
  /chats/{chatId}/messages/limit/{id}:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID чату.
        Повертає певну кількість повідомлень чату.
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: integer
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: return message ID
          schema:
            $ref: '#/definitions/messages.MessageListResponse'
        "500":
          description: get limit error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get limited message list
      tags:
      - message
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"